wip: fixing stuff
